Internal API

This page lists unexported functionality from Brillouin, that may be of interest to developers (but which is not exported and not covered by semver).

Unexported, internal functionality

Brillouin.BasisEnum — Type

BasisEnum <: Enum{Int8}

Enum type with instances

CARTESIAN
LATTICE

Used to indicate whether the coordinates of a KPath instance are referred to a lattice basis (LATTICE) or a cartesian basis (CARTESIAN).

See also setting and set_setting!.

Bravais.cartesianize! — Method

cartesianize!(c::Cell{D})

Transform a Wigner-Seitz cell c in a lattice basis to a Cartesian basis with (primitive) reciprocal basis vectors basis(c). Modifies c in-place.

Bravais.latticize! — Method

latticize!(c::Cell)

Transform a Wigner-Seitz cell c in a Cartesian basis to a lattice basis with (primitive) reciprocal lattice vectors basis(c). Modifies c in-place.

Bravais.latticize — Method

latticize(c::Cell{D}, basis::AbstractVector{<:AbstractVector{<:Real}})

Transform a Wigner-Seitz cell c in a Cartesian basis to a lattice basis with (primitive) reciprocal lattice vectors basis.

If c is not in a Cartesian basis (i.e., if setting(c) == LATTICE), c is returned as-is.

Brillouin.WignerSeitz.is_outward_oriented — Method

is_outward_oriented(c, n)

Return whether a face, specified by its center c and normal n, is outward pointing, assuming the face is part of a convex hull centered around origo.

Brillouin.KPaths.get_points_2d — Method

get_points_2d(ext_bt, Rs)

Return the labels and points in the 2D k-path associated with an extended Bravais type ext_bt and a (conventional) direct basis Rs (can be nothing if there is no dependence on the basis) as a Dict{Symbol, SVector{2, Float64}}.

Brillouin.KPaths.get_points_3d — Method

get_points_3d(ext_bt, Rs)

Return the labels and points in the 3D k-path associated with an extended Bravais type ext_bt and a (conventional) direct basis Rs (can be nothing if there is no dependence on the basis) as a Dict{Symbol, SVector{3, Float64}}.