API

Exported types

Exported methods

Brillouin.WignerSeitz.wignerseitz — Method

wignerseitz(Vs::AbstractVector{<:SVector{D}}; merge::Bool = true, Nmax = 3)
wignerseitz(Vs::AbstractVector{AbstractVector}; merge::Bool = true, Nmax = 3)
                                                            --> Cell{D}

Return a Cell{D} structure, containing the vertices and associated (outward oriented) faces, of the Wigner-Seitz cell defined by a basis Vs in D dimensions.

Keyword arguments

merge (default, true): if :true, co-planar faces are merged to form polygonal planar faces (e.g., triangles, quadrilaterals, and ngons generally). If false, raw "unprocessed" triangles (D=3) and segments (D=2) are returned instead. merge has no impact for D=1.

Nmax (default, 3): includes -Nmax:Nmax points in the initial lattice used to generate the underlying Voronoi tesselation. It is unwise to set this to anything lower than 3 without explicitly testing convergence; and probably unnecessary to increase it beyond 3 as well.

Brillouin.KPaths.cartesianize! — Method

cartesianize(kp::KPath, Gs::BasisLike)

Transform a k-path kp to a Cartesian coordinate system using a primitive basis Gs. Modifies the underlying dictionary in kp in-place.

Brillouin.KPaths.interpolate — Method

interpolate(kvs::AbstractVector{<:AbstractVector{<:Real}}, N::Integer) 
    --> Vector{<:Vector{<:Real}}, Int64

Return an interpolated k-path between discrete k-points in kvs, with approximately N interpolation points in total (typically fewer).

Since the actual number of interpolation points is not be guaranteed to equal N, the actual number of interpolation points is returned in addition to the interpolation itself.

Note that, in general, it is not possible to do this so that all interpolated k-points are equidistant; samples are however exactly equidistant across each linear segment defined by points in kvs and approximately equidistant across all segments.

See also splice.

Brillouin.KPaths.interpolate — Method

interpolate(kp::KPath, N::Integer; legacy::Bool=false)

Return an interpolation of kp with approximately N points distributed across all path segments. See also interpolate(::AbstractVector, ::Integer)].

Brillouin.KPaths.irrfbz_path — Function

irrfbz_path(sgnum::Integer, Rs::Union{Nothing, AbstractArray{var"#s12",1} where var"#s12"<:(StaticArrays.SArray{Tuple{D},var"#s13",1,D} where var"#s13"<:Real) where D}=nothing)   -->  ::KPath

Returns a k-path (::KPath) in the (primitive) irreducible Brillouin zone that includes all distinct high-symmetry lines and points as well as parts of the Brillouin zone boundary.

Rs refers to the direct basis of the conventional unit cell. For some space groups, it is needed to disambiguate the "extended" Bravais types that may differ depending on the lengths of the lattice vectors (because the Brillouin zone may depend on these lengths). If the requested space group is known to not fall under this case, Rs can be supplied as nothing (default).

Note that the returned k-points are given in the basis of the primitive reciprocal basis (see cartesianize!).

To interpolate the resulting KPath, see interpolate(::KPath, ::Integer) and splice(::KPath, ::Integer).

Data and referencing

All data is sourced from the SeeK HPKOT publication[1]: please cite the original work.

All paths currently assume time-reversal symmetry (or, equivalently, inversion symmetry), corresponding to the SeeK's [with inversion] setting. If neither inversion nor time-reversal, include the "inverted" **-k* paths as well manually.

[1] Hinuma, Pizzi, Kumagai, Oba, & Tanaka, Band structure diagram paths based on crystallography, Comp. Mat. Sci. 128, 140 (2017)

Brillouin.KPaths.paths — Method

paths(kp::KPath) -> Dict{Symbol, SVector{D,Float64}}

Return a vector of vectors, with each vector describing a connected path between between k-points referenced in kp (see also points(::KPath)).

Brillouin.KPaths.points — Method

points(kp::KPath) -> Dict{Symbol, SVector{D,Float64}}

Return a dictionary of the k-points (values) and associated k-labels (keys) referenced in kp.

Brillouin.KPaths.splice — Method

splice(kvs::AbstractVector{<:AbstractVector{<:Real}}, N::Integer) 
    --> Vector{<:Vector{<:Real}}

Return an interpolated k-path between the discrete k-points in kvs, with N interpolation points inserted in each segment defined by pairs of adjacent k-points.

See also interpolate.

Brillouin.KPaths.splice — Method

splice(kp::KPath, N::Integer)

Return an interpolation of kp with N points inserted into each segment of paths in kp. See also splice(::AbstractVector, ::Integer).

Exported constants