API

Types

Brillouin.WignerSeitz.Cell — Type

struct Cell{D} <: AbstractArray{Array{StaticArrays.SArray{Tuple{D}, Float64, 1, D}, 1}, 1}

verts::Array{StaticArrays.SVector{D, Float64}, 1} where D

faces::Vector{Vector{Int64}}

basis::StaticArrays.SArray{Tuple{D}, StaticArrays.SVector{D, Float64}, 1, D} where D

setting::Ref{Brillouin.BasisEnum}

Brillouin.KPaths.KPath — Type

struct KPath{D} <: Brillouin.KPaths.AbstractPath{Pair{Symbol, StaticArrays.SArray{Tuple{D}, Float64, 1, D}}}

points::Dict{Symbol, StaticArrays.SVector{D, Float64}} where D

paths::Vector{Vector{Symbol}}

basis::StaticArrays.SArray{Tuple{D}, StaticArrays.SVector{D, Float64}, 1, D} where D

setting::Ref{Brillouin.BasisEnum}

Brillouin.KPaths.KPathInterpolant — Type

struct KPathInterpolant{D} <: Brillouin.KPaths.AbstractPath{StaticArrays.SArray{Tuple{D}, Float64, 1, D}}

kpaths::Array{Array{StaticArrays.SVector{D, Float64}, 1}, 1} where D

labels::Vector{Dict{Int64, Symbol}}

basis::StaticArrays.SArray{Tuple{D}, StaticArrays.SVector{D, Float64}, 1, D} where D

setting::Ref{Brillouin.BasisEnum}

Exported methods

Brillouin.WignerSeitz.wignerseitz — Method

wignerseitz(basis::AbstractVector{<:SVector{D}}; merge::Bool = true, Nmax = 3)
wignerseitz(basis::AbstractVector{AbstractVector}; merge::Bool = true, Nmax = 3)
                                                            --> Cell{D}

Given a provided basis, return a Cell{D} containing the vertices and associated (outward oriented) faces of the Wigner-Seitz cell defined by basis in D dimensions. The returned vertices are given in the the basis of basis (see cartesianize! for conversion).

Keyword arguments

merge (default, true): if true, co-planar faces are merged to form polygonal planar faces (e.g., triangles, quadrilaterals, and ngons generally). If false, raw "unprocessed" triangles (D=3) and segments (D=2) are returned instead. merge has no impact for D=1.

Nmax (default, 3): includes -Nmax:Nmax points in the initial lattice used to generate the underlying Voronoi tesselation. It is unwise to set this to anything lower than 3 without explicitly testing convergence; and probably unnecessary to increase it beyond 3 as well.

Brillouin.KPaths.interpolate — Method

interpolate(kvs::AbstractVector{<:AbstractVector{<:Real}}, N::Integer)
    --> Vector{<:Vector{<:Real}}

Return an interpolated k-path between discrete k-points in kvs, with approximately N interpolation points in total (typically fewer).

Note that, in general, it is not possible to do this so that all interpolated k-points are equidistant; samples are however exactly equidistant across each linear segment defined by points in kvs and approximately equidistant across all segments.

See also interpolate(::KPath, ::Integer) and splice.

Brillouin.KPaths.interpolate — Method

interpolate(kp::KPath, length::Integer)
interpolate(kp::KPath; length::Integer, density::Real) --> KPathInterpolant

Return an interpolant of kp with approximately length points distributed approximately equidistantly across the full k-path (equidistance is measured in a Cartesian metric).

Note that the interpolant may contain fewer or more points than length (typically fewer). length can also be provided as a keyword argument.

As an alternative to specifying the desired total number of interpolate points via length, a desired density per unit (reciprocal) length can be specified via the keyword argument density.

See also interpolate(::AbstractVector{::AbstractVector{<:Real}}, ::Integer).

Brillouin.KPaths.irrfbz_path — Method

irrfbz_path(sgnum::Integer, Rs, [::Union{Val(D), Integer},]=Val(3))  -->  ::KPath{D}

Returns a k-path (::KPath) in the (primitive) irreducible Brillouin zone for a space group with number sgnum, (conventional) direct lattice vectors Rs, and dimension D. The path includes all distinct high-symmetry lines and points as well as relevant parts of the Brillouin zone boundary.

The dimension D (1, 2, or 3) is specified as the third input argument, preferably as a static Val{D} type parameter (or, type-unstably, as an <:Integer). Defaults to Val(3).

Rs refers to the direct basis of the conventional unit cell, i.e., not the primitive direct basis vectors. The setting of Rs must agree with the conventional setting choices in the International Tables of Crystallography. If Rs is a subtype of a StaticVector or NTuple, the dimension can be inferred from its (static) size; in this case, this dimension will take precedence (i.e. override, if different) over any dimension specified in the third input argument.

Notes

The returned k-points are given in the basis of the primitive reciprocal basis (see cartesianize!).

To interpolate a KPath, see interpolate(::KPath, ::Integer) and splice(::KPath, ::Integer).

All paths currently assume time-reversal symmetry (or, equivalently, inversion symmetry). If neither are present, include the "inverted" -k paths manually.

Data and referencing

3D paths are sourced from the SeeK HPKOT publication: please cite the original work [1].

[1] Hinuma, Pizzi, Kumagai, Oba, & Tanaka, Band structure diagram paths based on crystallography, Comp. Mat. Sci. 128, 140 (2017)

Brillouin.KPaths.paths — Method

paths(kp::KPath) -> Vector{Vector{Symbol}}

Return a vector of vectors, with each vector describing a connected path between between k-points referenced in kp (see also points(::KPath)).

Brillouin.KPaths.points — Method

points(kp::KPath{D}) -> Dict{Symbol, SVector{D,Float64}}

Return a dictionary of the k-points (values) and associated k-labels (keys) referenced in kp.

Brillouin.KPaths.splice — Method

splice(kvs::AbstractVector{<:AbstractVector{<:Real}}, N::Integer)
    --> Vector{<:Vector{<:Real}}

Return an interpolated k-path between the discrete k-points in kvs, with N interpolation points inserted in each segment defined by pairs of adjacent k-points.

See also splice(::KPath, ::Integer) and interpolate.

Brillouin.KPaths.splice — Method

splice(kp::KPath, N::Integer) --> KPathInterpolant

Return an interpolant of kp with N points inserted into each k-path segment of kp.

See also splice(::AbstractVector{::AbstractVector{<:Real}}, ::Integer).