API

Exported types

Exported methods

Brillouin.WignerSeitz.wignerseitz — Method

wignerseitz(Vs::AbstractVector{<:SVector{D}}; merge::Bool = true, Nmax = 3)
wignerseitz(Vs::AbstractVector{AbstractVector}; merge::Bool = true, Nmax = 3)
                                                            --> Cell{D}

Return a Cell{D} structure, containing the vertices and associated (outward oriented) faces, of the Wigner-Seitz cell defined by a basis Vs in D dimensions.

Keyword arguments

merge (default, true): if :true, co-planar faces are merged to form polygonal planar faces (e.g., triangles, quadrilaterals, and ngons generally). If false, raw "unprocessed" triangles (D=3) and segments (D=2) are returned instead. merge has no impact for D=1.

Nmax (default, 3): includes -Nmax:Nmax points in the initial lattice used to generate the underlying Voronoi tesselation. It is unwise to set this to anything lower than 3 without explicitly testing convergence; and probably unnecessary to increase it beyond 3 as well.

Brillouin.KPath.interpolate_path — Method

interpolate_path(kvs::AbstractVector{<:AbstractVector{<:Real}}, N::Integer) 
    --> Vector{<:Vector{<:Real}}, Int64

Return an interpolated $k$-path between discrete $k$-points in kvs, with approximately N interpolation points in total (typically fewer).

Since the actual number of interpolation points is not be guaranteed to equal N, the actual number of interpolation points is returned in addition to the interpolation itself.

Note that, in general, it is not possible to do this so that all interpolated $k$-points are equidistant; samples are however exactly equidistant across each linear segment defined by points in kvs and approximately equidistant across all segments.

See also splice_path.

Brillouin.KPath.irrfbz_path — Method

irrfbz_path(sgnum::Integer, Nk::Integer, Rs::Union{Nothing, AbstractArray{StaticArrays.SArray{Tuple{D},var"#s12",1,D} where var"#s12"<:Real,1} where D}=nothing; 
    pathtype::String="SeeK", has_inversion_or_tr::Bool=true,
    splice::Bool=false, legacy::Bool=false)
                                            --> paths_kvs, paths_labs, lab2kv

Returns a k-path in the (primitive) irreducible Brillouin zone that includes all distinct high-symmetry lines and points as well as parts of the Brillouin zone boundary.

Rs refers to the direct basis of the conventional unitcell. For some space groups, it is needed to disambiguate the "extended" Bravais types that may differ depending on the lengths of the lattice vectors (because the Brillouin zone may depend on these lengths). If the requested space group is known to not fall under this case, Rs can be supplied as nothing (default).

Data and referencing

All data is sourced from the SeeK publication[1]: please cite the original work.

All paths currently assume time-reversal symmetry (or, equivalently, inversion symmetry), corresponding to the SeeK's [with inversion] setting. If neither inversion nor time-reversal symmetry is present, additional paths may be required (SeeK's [no inversion] setting).

[1] Hinuma, Pizzi, Kumagai, Oba, & Tanaka, Band structure diagram paths based on crystallography, Comp. Mat. Sci. 128, 140 (2017)](http://dx.doi.org/10.1016/j.commatsci.2016.10.015) (see also online interface at https://www.materialscloud.org/work/tools/seekpath).

Brillouin.KPath.splice_path — Method

splice_path(kvs::AbstractVector{<:AbstractVector{<:Real}}, N::Integer) 
    --> Vector{<:Vector{<:Real}}

Computes an interpolated $k$-path between the discrete $k$-points in kvs, with N interpolation points inserted in each segment defined by pairs of adjacent $k$-points.

See also interpolate_path.

Exported constants