API

Types

Brillouin.WignerSeitz.Cell — Type

struct Cell{D} <: AbstractArray{Array{StaticArrays.SArray{Tuple{D}, Float64, 1, D}, 1}, 1}

verts::Array{StaticArrays.SVector{D, Float64}, 1} where D

faces::Vector{Vector{Int64}}

basis::StaticArrays.SArray{Tuple{D}, StaticArrays.SVector{D, Float64}, 1, D} where D

Brillouin.KPaths.KPath — Type

struct KPath{D} <: Brillouin.KPaths.AbstractPath{Pair{Symbol, StaticArrays.SArray{Tuple{D}, Float64, 1, D}}}

points::Dict{Symbol, StaticArrays.SVector{D, Float64}} where D

paths::Vector{Vector{Symbol}}

Brillouin.KPaths.KPathInterpolant — Type

struct KPathInterpolant{D} <: Brillouin.KPaths.AbstractPath{StaticArrays.SArray{Tuple{D}, Float64, 1, D}}

kpaths::Array{Array{StaticArrays.SVector{D, Float64}, 1}, 1} where D

labels::Vector{Dict{Int64, Symbol}}

Exported methods

Brillouin.WignerSeitz.wignerseitz — Method

wignerseitz(Vs::AbstractVector{<:SVector{D}}; merge::Bool = true, Nmax = 3)
wignerseitz(Vs::AbstractVector{AbstractVector}; merge::Bool = true, Nmax = 3)
                                                            --> Cell{D}

Return a Cell{D} structure, containing the vertices and associated (outward oriented) faces, of the Wigner-Seitz cell defined by a basis Vs in D dimensions.

Keyword arguments

merge (default, true): if true, co-planar faces are merged to form polygonal planar faces (e.g., triangles, quadrilaterals, and ngons generally). If false, raw "unprocessed" triangles (D=3) and segments (D=2) are returned instead. merge has no impact for D=1.

Nmax (default, 3): includes -Nmax:Nmax points in the initial lattice used to generate the underlying Voronoi tesselation. It is unwise to set this to anything lower than 3 without explicitly testing convergence; and probably unnecessary to increase it beyond 3 as well.

Brillouin.KPaths.cartesianize! — Method

cartesianize!(kpi::KPathInterpolant{D}, Gs::BasisLike)

Transform an interpolated k-path kpi in a lattice basis to a Cartesian basis with (primitive) reciprocal lattice vectors Gs. Modifies kpi in-place.

Warning

If kpi was originally created from a KPath in a non-cartesian basis, its points will not be equidistantly spaced when converted to a Cartesian basis. In particular, for a kp::KPath in the lattice basis, it is generally true that interpolate(cartesianize(kp, pGs), N) ≠ cartesianize(interpolate(kp, N), pGs)), unless the reciprocal basis pGs is square. Only interpolate(cartesianize(kp, pGs) is guaranteed to produce paths that are approximately equidistantly spaced.

It is therefore strongly recommended not to use cartesianize!(::KPathInterpolant) except in combination with latticize!(::KPathInterpolant).

Brillouin.KPaths.cartesianize! — Method

cartesianize!(kp::KPath{D}, Gs::BasisLike{D})

Transform a k-path kp in a lattice basis to a Cartesian basis with (primitive) basis vectors Gs. Modifies kp in-place.

Brillouin.KPaths.interpolate — Method

interpolate(kvs::AbstractVector{<:AbstractVector{<:Real}}, N::Integer)
    --> Vector{<:Vector{<:Real}}

Return an interpolated k-path between discrete k-points in kvs, with approximately N interpolation points in total (typically fewer).

Note that, in general, it is not possible to do this so that all interpolated k-points are equidistant; samples are however exactly equidistant across each linear segment defined by points in kvs and approximately equidistant across all segments.

See also interpolate(::KPath, ::Integer) and splice.

Brillouin.KPaths.interpolate — Method

interpolate(kp::KPath, N::Integer) --> KPathInterpolant

Return an interpolant of kp with approximately N points distributed approximately equidistantly across the full k-path.

Note that the interpolant may contain fewer or more points than N (typically fewer).

See also interpolate(::AbstractVector{::AbstractVector{<:Real}}, ::Integer).

Brillouin.KPaths.irrfbz_path — Method

irrfbz_path(sgnum::Integer,
            Rs::Union{Nothing, AVec{<:AVec{<:Real}}}=nothing,
            [::Union{Val(D), Integer},]=Val(3))                     -->  ::KPath{D}

Returns a k-path (::KPath) in the (primitive) irreducible Brillouin zone that includes all distinct high-symmetry lines and points as well as relevant parts of the Brillouin zone boundary.

The dimension D (1, 2, or 3) is specified as the third input argument, preferably as a static Val{D} type parameter (or, type-unstably, as an <:Integer). Defaults to Val(3).

Rs refers to the direct basis of the conventional unit cell. For some space groups, it is needed to disambiguate the "extended" Bravais types that may differ depending on the lengths of the lattice vectors (because the Brillouin zone may depend on these lengths). If the requested space group is known to not fall under this case, Rs can be supplied as nothing (default). If Rs is a subtype of a StaticVector, the dimension can be inferred from its (static) size; no explicit input of dimension is needed in this case.

Notes

The returned k-points are given in the basis of the primitive reciprocal basis (see cartesianize!).

To interpolate a KPath, see interpolate(::KPath, ::Integer) and splice(::KPath, ::Integer).

All paths currently assume time-reversal symmetry (or, equivalently, inversion symmetry). If neither are present, include the "inverted" -k paths manually.

Data and referencing

3D paths are sourced from the SeeK HPKOT publication: please cite the original work [1].

[1] Hinuma, Pizzi, Kumagai, Oba, & Tanaka, Band structure diagram paths based on crystallography, Comp. Mat. Sci. 128, 140 (2017)

Brillouin.KPaths.latticize! — Method

latticize!(kpi::KPathInterpolant{D}, Gs::BasisLike)

Transform an interpolated k-path kpi in a Cartesian basis to a lattice basis with (primitive) reciprocal lattice vectors Gs. Modifies kpi in-place.

Brillouin.KPaths.latticize! — Method

latticize!(kp::KPath{D}, Gs::BasisLike{D})

Transform a k-path kpi in a Cartesian basis to a lattice basis with (primitive) reciprocal lattice vectors Gs. Modifies kp in-place.

Brillouin.KPaths.paths — Method

paths(kp::KPath) -> Vector{Vector{Symbol}}

Return a vector of vectors, with each vector describing a connected path between between k-points referenced in kp (see also points(::KPath)).

Brillouin.KPaths.points — Method

points(kp::KPath{D}) -> Dict{Symbol, SVector{D,Float64}}

Return a dictionary of the k-points (values) and associated k-labels (keys) referenced in kp.

Brillouin.KPaths.splice — Method

splice(kvs::AbstractVector{<:AbstractVector{<:Real}}, N::Integer)
    --> Vector{<:Vector{<:Real}}

Return an interpolated k-path between the discrete k-points in kvs, with N interpolation points inserted in each segment defined by pairs of adjacent k-points.

See also splice(::KPath, ::Integer) and interpolate.

Brillouin.KPaths.splice — Method

splice(kp::KPath, N::Integer) --> KPathInterpolant

Return an interpolant of kp with N points inserted into each k-path segment of kp.

See also splice(::AbstractVector{::AbstractVector{<:Real}}, ::Integer).