k-space paths

Default paths

To generate a k-path for, say, the space group of diamond (space group 227; a cubic face-centered Bravais lattice), we can call irrfbz_path, which will return a minimal path in the irreducible Brillouin zone:

using Brillouin
sgnum = 227
Rs = [[1,0,0], [0,1,0], [0,0,1]] # conventional direct basis
kp = irrfbz_path(sgnum, Rs)

KPath{3} (6 points, 2 paths, 8 points in paths):
 points: :U => [0.625, 0.25, 0.625]
         :W => [0.5, 0.25, 0.75]
         :K => [0.375, 0.375, 0.75]
         :Γ => [0.0, 0.0, 0.0]
         :L => [0.5, 0.5, 0.5]
         :X => [0.5, 0.0, 0.5]
  paths: [:Γ, :X, :U]
         [:K, :Γ, :L, :W, :X]

The path data is sourced from the HPKOT paper (or, equivalently, the SeeK-path Python package).

The resulting KPath structure initially gives the k-point coordinates in the basis of the primitive reciprocal basis. To convert to a Cartesian basis, we use cartesianize!:

pGs = 2π.*[[-1,1,1], [1,-1,1], [1,1,-1]] # primitive reciprocal basis
cartesianize!(kp, pGs)

KPath{3} (6 points, 2 paths, 8 points in paths):
 points: :U => [1.5707963267948966, 6.283185307179586, 1.5707963267948966]
         :W => [3.141592653589793, 6.283185307179586, 0.0]
         :K => [4.71238898038469, 4.71238898038469, 0.0]
         :Γ => [0.0, 0.0, 0.0]
         :L => [3.141592653589793, 3.141592653589793, 3.141592653589793]
         :X => [0.0, 6.283185307179586, 0.0]
  paths: [:Γ, :X, :U]
         [:K, :Γ, :L, :W, :X]

We can visualize the k-path using PlotlyJS.jl:

using PlotlyJS
Pᵏ = plot(kp)

Usually, it'll be more helpful to understand the path's geometry in the context of the associated Brillouin zone. To visualize this, we can plot the combination of a Cell (created via wignerseitz) and a KPath:

c = wignerseitz(pGs)
Pᶜ⁺ᵏ = plot(c, kp)

Interpolation

Interpolation of a KPath structure can be achieved using either interpolate(::KPath, ::Integer) or splice(::KPath, ::Integer), returning a KPathInterpolant. As an example, interpolate(kp, N) returns an interpolation with a target of N interpolation points, distributed as equidistantly as possible:

kpi = interpolate(kp, 100)

99-element KPathInterpolant{3}:
 [0.0, 0.0, 0.0]
 [0.0, 0.28559933214452665, 0.0]
 [0.0, 0.5711986642890533, 0.0]
 [0.0, 0.8567979964335799, 0.0]
 [0.0, 1.1423973285781066, 0.0]
 [0.0, 1.427996660722633, 0.0]
 [0.0, 1.7135959928671598, 0.0]
 [0.0, 1.9991953250116865, 0.0]
 [0.0, 2.284794657156213, 0.0]
 [0.0, 2.5703939893007397, 0.0]
 ⋮
 [2.284794657156213, 6.283185307179586, 0.0]
 [1.9991953250116865, 6.283185307179586, 0.0]
 [1.71359599286716, 6.283185307179586, 0.0]
 [1.4279966607226333, 6.283185307179586, 0.0]
 [1.1423973285781068, 6.283185307179586, 0.0]
 [0.8567979964335799, 6.283185307179586, 0.0]
 [0.5711986642890534, 6.283185307179586, 0.0]
 [0.28559933214452676, 6.283185307179586, 0.0]
 [0.0, 6.283185307179586, 0.0]

The returned KPathInterpolant implements the AbstractVector interface, with iterants returning SVector{D, Float64} elements. To get a conventional "flat" vector, we can simply call collect(kpi).

Internally, KPathInterpolantincludes additional structure and information: namely, the high-symmetry points and associated labels along the path as well as a partitioning into connected vs. disconnected path segments.

Band structure

The additional structure in KPathInterpolations enable convenient and clear visualization of band structure diagrams in combination with PlotlyJS.

To illustrate this, suppose we are considering a tight-binding problem for an s-orbital situated at the 1a Wyckoff position. Such a problem has a single band with dispersion [1] (assuming a cubic side length $a = 1$):

\[\epsilon(\mathbf{k}) = 4\gamma\Bigl( \cos \tfrac{1}{2}k_x \cos \tfrac{1}{2}k_y + \cos \tfrac{1}{2}k_y \cos \tfrac{1}{2}k_z + \cos \tfrac{1}{2}k_z \cos \tfrac{1}{2}k_x \Bigr)\]

We can calculate the associated energy band along our earlier k-path interpolation kpi easily:

function ϵ(k; γ::Real=1.0)
    4γ * (cos(k[1]/2)*cos(k[2]/2) + cos(k[2]/2)*cos(k[3]/2) + cos(k[3]/2)*cos(k[1]/2))
end
band = ϵ.(kpi)

99-element Vector{Float64}:
 12.0
 11.918571535047462
 11.67594378891598
 11.277055962836148
 10.73002826264945
 10.045996594834067
  9.238885871562282
  8.325126539644781
  7.3233201040150915
  6.253860454731438
  ⋮
 -4.0
 -4.0
 -4.0
 -3.9999999999999996
 -3.9999999999999996
 -4.0
 -4.0
 -4.0
 -4.0

And we can then visualize the associated band via an overloaded PlotlyJS plot call:

P = plot(kpi, [band])

If we have multiple bands, say $\epsilon_1(\mathbf{k}) = \epsilon(\mathbf{k})$ and $\epsilon_2(\mathbf{k}) = 20 - \tfrac{1}{2}\epsilon(\mathbf{k})$, we can easily plot that by collecting the bands in a single vector (or concatenating into a matrix):

band1 = ϵ.(kpi)
band2 = 20 .- (1/2).*band1
P¹² = plot(kpi, [band1, band2])

PlotlyJS.plot — Method

plot(kpi::KPathInterpolant, bands, [layout]; kwargs...)

Plot a dispersion diagram for provided bands and k-path interpolant kpi.

bands must be an iterable of iterables of <:Reals (e.g., a Vector{Vector{Float64}}), with the first iteration running over distinct energy bands, and the second running over distinct k-points in kpi. Note that the length of each iterant of bands must equal length(kpi).

Alternatively, bands can be an AbstractMatrix{<:Real}, with columns interpreted as distinct energy bands and rows as distinct k-points.

Keyword arguments kwargs

ylims: y-axis limits (default: quasi-tight around bands's range)

ylabel: y-axis label (default: "Energy")

[1] See e.g. http://www.physics.rutgers.edu/~eandrei/chengdu/reading/tight-binding.pdf